# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vehicle Transport Automation",
    "description": "Control Panel for Email-to-ClickUp Pipeline with ML Training Support",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/models/versions": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List Model Versions",
        "description": "List model versions with optional filtering.",
        "operationId": "list_model_versions_api_models_versions_get",
        "parameters": [
          {
            "name": "auction_type_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Auction Type Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/versions/active/{auction_type_id}": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get Active Model",
        "description": "Get the active model version for an auction type.",
        "operationId": "get_active_model_api_models_versions_active__auction_type_id__get",
        "parameters": [
          {
            "name": "auction_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Auction Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/versions/{model_id}/promote": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Promote Model",
        "description": "Promote a model version to active status.\n\nThis deactivates any other active model for the same auction type.",
        "operationId": "promote_model_api_models_versions__model_id__promote_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/versions/{model_id}": {
      "delete": {
        "tags": [
          "Models"
        ],
        "summary": "Archive Model",
        "description": "Archive a model version (soft delete).",
        "operationId": "archive_model_api_models_versions__model_id__delete",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/train": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Start Training",
        "description": "Start a new training job.\n\nThis will train a PEFT/LoRA adapter for the specified auction type\nusing the collected training examples.",
        "operationId": "start_training_api_models_train_post",
        "parameters": [
          {
            "name": "sync",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Run synchronously (wait for completion)",
              "default": false,
              "title": "Sync"
            },
            "description": "Run synchronously (wait for completion)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingJobRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/jobs": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "List Training Jobs",
        "description": "List training jobs with optional filtering.",
        "operationId": "list_training_jobs_api_models_jobs_get",
        "parameters": [
          {
            "name": "auction_type_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Auction Type Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/jobs/{job_id}": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Get Training Job",
        "description": "Get details of a training job.",
        "operationId": "get_training_job_api_models_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Cancel Training Job",
        "description": "Cancel a running training job.",
        "operationId": "cancel_training_job_api_models_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelVersionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelVersionResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ModelVersionListResponse",
        "description": "List of model versions."
      },
...
      "TrainingJobRequest": {
        "properties": {
          "auction_type_id": {
            "type": "integer",
            "title": "Auction Type Id",
            "description": "Auction type to train for"
          },
          "version_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Tag",
            "description": "Version tag (auto-generated if not provided)"
          },
          "base_model": {
            "type": "string",
            "title": "Base Model",
            "description": "Base model to fine-tune",
            "default": "microsoft/layoutlmv3-base"
          },
          "epochs": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Epochs",
            "description": "Training epochs",
            "default": 3
          },
          "learning_rate": {
            "type": "number",
            "title": "Learning Rate",
            "description": "Learning rate",
            "default": 2e-05
          },
          "batch_size": {
            "type": "integer",
            "maximum": 32.0,
            "minimum": 1.0,
            "title": "Batch Size",
            "description": "Batch size",
            "default": 4
          }
        },
        "type": "object",
        "required": [
          "auction_type_id"
        ],
        "title": "TrainingJobRequest",
        "description": "Request to start a training job."
      },
...
    }
  }
}